rename cloud-tools -> pligos
