+++ go.mod
@@ -1,4 +1,4 @@
-module realcloud.tech/cloud-tools
+module realcloud.tech/pligos
 
 require (
 	github.com/BurntSushi/toml v0.3.1 // indirect
